+++ src/modules/orders/repositories/orderRepository.js
@@ -29,14 +29,16 @@ async function deleteDatabase(collection, filter) {
 async function orderRepository(fastify) {
   const collection = fastify.mongo.db.collection('orders');
 
-  fastify.decorate('orderRepository', {
-    findByIdProperty: (idProperty) => findByIdProperty(collection, idProperty),
-    findByQuery: (query) => findByQuery(collection, query),
-    updateDatabase: (filter, data) => updateDatabase(collection, filter, data),
-    insertDatabase: (data) => insertDatabase(collection, data),
-    deleteDatabase: (filter) => deleteDatabase(collection, filter),
-  });
-}
+  if (!fastify.hasDecorator('orderRepository')) {
+    fastify.decorate('orderRepository', {
+      findByIdProperty: (idProperty) => findByIdProperty(collection, idProperty),
+      findByQuery: (query) => findByQuery(collection, query),
+      updateDatabase: (filter, data) => updateDatabase(collection, filter, data),
+      insertDatabase: (data) => insertDatabase(collection, data),
+      deleteDatabase: (filter) => deleteDatabase(collection, filter),
+    });
+  }
+  }
 
 export default fp(orderRepository, {
   name: 'orderRepository',
